backend/requirements.txt
anyio==3.6.1
asgiref==3.5.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
DateTime==4.4
Deprecated==1.2.13
fastapi==0.78.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
pycparser==2.21
pydantic==1.9.1
PyGithub==1.55
PyJWT==2.4.0
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.19.1
typing-extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
wrapt==1.14.1
zope.interface==5.4.0
